Moving the httpPlayback code in to it's own Trait
