+++ src/API/NTLMSoapClient.php
@@ -5,10 +5,7 @@ namespace jamesiarmes\PEWS\API;
 use SoapClient;
 use GuzzleHttp;
 use SoapHeader;
-use GuzzleHttp\HandlerStack;
-use GuzzleHttp\Middleware;
-use GuzzleHttp\Handler\MockHandler;
-use GuzzleHttp\Psr7\Response;
+use jamesiarmes\PEWS\HttpPlayback\HttpPlayback;
 
 /**
  * Contains NTLMSoapClient.
@@ -42,6 +39,7 @@ use GuzzleHttp\Psr7\Response;
  */
 class NTLMSoapClient extends SoapClient
 {
+    use HttpPlayback;
     /**
      * Username for authentication on the exchnage server
      *
@@ -65,17 +63,7 @@ class NTLMSoapClient extends SoapClient
 
     protected $__last_request_headers;
 
-    protected static $mode = 'playback';
-
-    protected static $callList = [];
-
-    protected static $recordLocation;
-
-    protected $container = [];
-
     protected $_responseCode;
-
-    private static $shutdownRegistered = false;
 
     /**
      * @TODO: Make this smarter. It should know and search what headers to remove on what actions
@@ -158,67 +146,9 @@ class NTLMSoapClient extends SoapClient
             $this->setHttpClient($options['httpClient']);
         }
 
-        if (!self::$shutdownRegistered) {
-            register_shutdown_function(array($this, 'endRecord'));
-            self::$shutdownRegistered = true;
-        }
-
-        if ($options['httpPlayback']['mode'] != null) {
-            self::$mode = $options['httpPlayback']['mode'];
-        }
+        $this->setPlaybackOptions($options['httpPlayback']);
 
         parent::__construct($wsdl, $options);
-    }
-
-    /**
-     * @var GuzzleHttp\Client
-     */
-    private static $client;
-
-    /**
-     * Get the client for making calls
-     *
-     * @return GuzzleHttp\Client
-     */
-    public function getHttpClient()
-    {
-        if (self::$client == null) {
-            $handler = HandlerStack::create();
-
-            if (self::$mode == 'record') {
-                $this->container = [];
-                $history = Middleware::history($this->container);
-                $handler->push($history);
-            } elseif (self::$mode == 'playback') {
-                $recordings = $this->getRecordings();
-
-                $playList = $recordings;
-                $mockedResponses = [];
-                foreach ($playList as $item) {
-                    $mockedResponses[] = new Response($item['statusCode'], $item['headers'], $item['body']);
-                }
-
-                $mockHandler = new MockHandler($mockedResponses);
-                $handler = HandlerStack::create($mockHandler);
-            }
-
-            self::$client = new GuzzleHttp\Client(['handler' => $handler]);
-        }
-
-        return self::$client;
-    }
-
-    /**
-     * Sets the client
-     *
-     * @param GuzzleHttp\Client $client
-     * @return $this
-     */
-    public function setHttpClient($client)
-    {
-        $this->client = $client;
-
-        return $this;
     }
 
     /**
@@ -253,8 +183,6 @@ class NTLMSoapClient extends SoapClient
 
         $this->__last_request_headers = $headers;
         $this->_responseCode = $response->getStatusCode();
-
-        $this->recordCalls();
 
         return $response->getBody()->__toString();
     }
@@ -292,55 +220,5 @@ class NTLMSoapClient extends SoapClient
     public function getResponseCode()
     {
         return $this->_responseCode;
-    }
-
-    public function recordCalls()
-    {
-        if (self::$mode != 'record') {
-            return;
-        }
-
-        foreach ($this->container as $item) {
-            /** @var Response $response */
-            $response = $item['response'];
-            self::$callList[] = [
-                'statusCode' => $response->getStatusCode(),
-                'headers' => $response->getHeaders(),
-                'body' => $response->getBody()->__toString()
-            ];
-        }
-
-        $this->container = [];
-    }
-
-    public static function getRecordLocation()
-    {
-        if (!self::$recordLocation) {
-            $dir = __DIR__;
-            $dirPos = strrpos($dir, "src/API");
-            $dir = substr($dir, 0, $dirPos);
-
-            self::$recordLocation = $dir . 'Resources/recordings/saveState.json';
-        }
-
-        return self::$recordLocation;
-    }
-
-    public static function getRecordings()
-    {
-        $saveLocation = self::getRecordLocation();
-        $records = json_decode(file_get_contents($saveLocation), true);
-
-        return $records;
-    }
-
-    public function endRecord()
-    {
-        if (self::$mode != 'record') {
-            return;
-        }
-
-        $saveLocation = self::getRecordLocation();
-        file_put_contents($saveLocation, json_encode(self::$callList));
     }
 }
